requirements.txt
python-telegram-bot[fastapi]==20.3
fastapi
uvicorn
requests
